pipelines/airflow/requirements.txt
kafka-python
requests
oauthlib
psycopg2-binary
geopy
firebase_admin